Use grid space calculations to compute highlights
Wraps the _getGridHighlightPositions method to return a custom set
based on grid spaces. For radius based templates, this set is the
highlight. For cones, the set is additionally filtered to check the
offset angle. If the angle is less than about one half of the angle,
then it is included.

TODO: Line templates